+++ halo2_proofs/src/plonk/permutation/verifier.rs
@@ -3,7 +3,7 @@ use group::ff::{Field, PrimeField};
 use std::iter;
 
 use super::super::{circuit::Any, ChallengeBeta, ChallengeGamma, ChallengeX};
-use super::{Argument, VerifyingKey};
+use super::{permutation_chunk_len, Argument, VerifyingKey};
 use crate::{
     arithmetic::CurveAffine,
     plonk::{self, Error},
@@ -40,7 +40,7 @@ impl Argument {
         vk: &plonk::VerifyingKey<C>,
         transcript: &mut T,
     ) -> Result<Committed<C>, Error> {
-        let chunk_len = vk.cs_degree - 2;
+        let chunk_len = permutation_chunk_len(vk.cs_degree);
 
         let permutation_product_commitments = self
             .columns
@@ -116,7 +116,7 @@ impl<C: CurveAffine> Evaluated<C> {
         gamma: ChallengeGamma<C>,
         x: ChallengeX<C>,
     ) -> impl Iterator<Item = C::Scalar> + 'a {
-        let chunk_len = vk.cs_degree - 2;
+        let chunk_len = permutation_chunk_len(vk.cs_degree);
         iter::empty()
             // Enforce only for the first set.
             // l_0(X) * (1 - z_0(X)) = 0